requirements.txt
astroid==3.2.2
autopep8==2.3.0
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
Flask==3.0.3
Flask-MQTT==1.2.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fonttools==4.53.0
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mccabe==0.7.0
msgspec==0.18.6
numpy==2.0.0
packaging==24.1
paho-mqtt==1.6.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pylint==3.2.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
tomlkit==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
